Use lower horizontal resolution to find the item box.
This is done by only checking for a black square every 20 px. Additionally,
after a black square is found, the rest of this row (in the item tooltip)
is skipped entirely. The reason for this is that we are looking for rows
that are black for their full width, so if we hit a row which is completely
black, every other point in the same row would lead to the same frame.
If we hit a row which isn't, we don't care about it anyway.
As a result, the performance is greatly improved.